Implemented Leave Validation (from Luke) Implemented Overtime Validation Updated apply & edit views to display validation errors
[Bug] Validation for LeaveBalance does not work because Employee is not binded by form. Commented out.

+++ src/main/java/sg/edu/nus/iss/lms/model/LeaveEntitlement.java
@@ -20,15 +20,15 @@ public class LeaveEntitlement {
 	@ManyToOne
 	private LeaveType leaveType;
 	
-	private Integer entitlement;
+	private double entitlement;
 	
-	private Integer leaveBalance;
+	private double leaveBalance;
 	
 	@ManyToOne
 	private Employee employee;
 	
 	// TestData Constructor
-	public LeaveEntitlement(Employee employee, LeaveType leaveType, Integer entitlement) {
+	public LeaveEntitlement(Employee employee, LeaveType leaveType, double entitlement) {
 		this.employee = employee;
 		this.leaveType = leaveType;
 		this.entitlement = entitlement;
